More efficient echo provider.
