--- VoiceCraft.Mobile/VoiceCraft.Mobile/Interfaces/IEffect.cs
@@ -1,7 +1,0 @@
-﻿namespace VoiceCraft.Mobile.Interfaces
-{
-    public interface IEffect
-    {
-        float ApplyEffect(float sample);
-    }
-}
