--- VoiceCraft.Windows/Interfaces/IEffect.cs
@@ -1,7 +1,0 @@
-﻿namespace VoiceCraft.Windows.Interfaces
-{
-    public interface IEffect
-    {
-        float ApplyEffect(float sample);
-    }
-}
